fix: 컨트롤러간 통신시 발생하는 오류 수정
- delegate 네이밍 수정
- 수정이 반영되지 않는 문제 수정

+++ CloudNotes/CloudNotes/Model/Memo.swift
@@ -10,7 +10,7 @@ import Foundation
 struct Memo: Codable {
     var title: String
     var body: String
-    var date: Int
+    var date: Double
     
     enum CodingKeys: String, CodingKey {
         case title
